Extract HTTP functionality to HttpClient for easier mocking
